fix expanded macro type error

+++ os/include/fs_misc.h
@@ -322,28 +322,28 @@ struct file_desc {
 //added by sundong 2023.5.26
 #define RD_BLOCK_SCHED(dev,block_nr,fsbuf) rw_blocks_sched(DEV_READ, \
 				       dev,				\
-				       (block_nr) * BLOCK_SIZE,		\
+				       (u64)(block_nr) * BLOCK_SIZE,		\
 				       BLOCK_SIZE, /* read one block */ \
 				       proc2pid(p_proc_current),/*current task id*/			\
 				       fsbuf);
 //added by sundong 2023.5.26
 #define WR_BLOCK_SCHED(dev,block_nr,fsbuf) rw_blocks_sched(DEV_WRITE, \
 				       dev,				\
-				       (block_nr) * BLOCK_SIZE,		\
+				       (u64)(block_nr) * BLOCK_SIZE,		\
 				       BLOCK_SIZE, /* write one block */ \
 				       proc2pid(p_proc_current),				\
 				       fsbuf);
 					
 #define RD_BLOCK(dev,block_nr,fsbuf) rw_blocks(DEV_READ, \
 				       dev,				\
-				       (block_nr) * BLOCK_SIZE,		\
+				       (u64)(block_nr) * BLOCK_SIZE,		\
 				       BLOCK_SIZE, /* read one block */ \
 				       proc2pid(p_proc_current),/*current task id*/			\
 				       fsbuf);
 //added by sundong 2023.5.26
 #define WR_BLOCK(dev,block_nr,fsbuf) rw_blocks(DEV_WRITE, \
 				       dev,				\
-				       (block_nr) * BLOCK_SIZE,		\
+				       (u64)(block_nr) * BLOCK_SIZE,		\
 				       BLOCK_SIZE, /* write one block */ \
 				       proc2pid(p_proc_current),				\
 				       fsbuf);
